combine all the q event tests
